Fixed PlexAPI installation when permission denied
I got an error when running outside a venv, this should fix it.

ERROR: Could not install packages due to an EnvironmentError: [Errno 13] Permission denied: '/usr/local/lib/python3.6/dist-packages/plexapi'
Consider using the `--user` option or check the permissions.

I only changed the subprocess call so it checks the output,
catch the error and run it again with "--user" instead.